Move ChangeKind to PatchSetAttribute
When adding the ChangeKind for JSON data, I added it into
PatchSetCreatedEvent which ended up being a pretty lousy place to
add it, since it has little to do with the event, and more to do
with the PatchSetAttribute itself.  This moves the data to the
PatchSet JSON structure, alongside the draft attribute.

Change-Id: Ia3791789af2ddb6efc70d78f38595519b5d6ab5f

Former-commit-id: 0c6bc94d0a6afea1c091114d4ab16825a2c9c5d2