Buck: Allow to co-exist in tree and standalone plugins
To allow standalone plugins to be built with Bucklets project we need
to be able to include the //bucklets/gerrit_plugin.bucklet from both
in tree and standalone build.  To achieve this create a dummy
bucklets/gerrit_plugin.bucklet file in Gerrit tree, as it doesn't suppose
to have any effects in tree build.

Change-Id: I885199a66cc71a756176f7edaca5f2cf0509e827

Former-commit-id: 44d0564a1ae1efc39de3ca5949d6c461523d4d80