Clarify the List Comments REST endpoint
The List Comments endpoint is a little confusing because users
expect this to get file as well as message comments.  This
endpoint only retrieves the former, not the latter.  Message
comments can be retrieved with the Get Change Detail endpoint.
This change just updates the docs to be more clear.

Issue: 2693
Change-Id: Id4fcbcaa0e755cc650cbaaf07b57dbae7e4a370f

Former-commit-id: 23845a1e6f3be4c7b059eb64edcd524265721fe8