Add support for project aware TopMenu extensions
Currently TopMenu extensions are not context aware, which is not a
problem for most of them, but in case of the 'Projects' section this
could be problematic.

With this approach plugins cannot contribute TopMenu items that will
only be shown when a project is selected and the url (and the history
item) contains a project name.

This change improves this by adding a possibility to provide a menu
item within the 'Projects' context without the previously mentioned
limitation.

Now when a TopMenu item contains a '${projectName}' placeholder it
will be replaced with name of project that is selected in the
'Projects' section.

Change-Id: I516f129dc73b5b98fd6c2d385690e51febfc00a3
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>

Former-commit-id: e8de74ff2e8693b1b3921b783f3275ff17304d60