Buck: Fix plugin build by not merging manifests from dependent JARs
After posting this mail to mailing list [1], filing this issue [2],
and creating this pull request [3], Buck team provided a solution to
the problem: when multiple JARs contain MANIFEST files, the default
behaviour of java_binary() was to merge the content of those manifest
files together. That breaks our plugins, as the content from other JARs
interfere with the manifest content of plugins.

Recent Buck version [4] provides undocumented parameter to change this
behaviour.

[1] https://groups.google.com/forum/#!topic/buck-build/M7Nu71okdYY
[2] https://github.com/facebook/buck/issues/86
[3] https://github.com/facebook/buck/pull/87
[4] https://github.com/facebook/buck/commit/\
96ba1967ef941558bc4c12ce88d5eaeba0b23275

Change-Id: Ia228fb51cfe8db678c2551b9cd1802bdaf4d44fc

Former-commit-id: 9d6a453d405f543f19408bef6c4376034cbc5b97