Add ChangeMessagesUtil to assist with notedb migration
Like ApprovalsUtil, we added an abstraction layer to assist in the
migration to notedb. This utility class helps to write to either the
notedb or the reviewdb depending on the state of the NotesMigration
instance. Additionally, in this change, I modified all callers of
ChangeMessageAccess (which uses the ReviewDb) to instead use
ChangeMessagesUtil.

Change-Id: Id007ed8e01bd1318f70d804124e1f482a77234fd

Former-commit-id: 086006ec35ef2a5bddb14a3a4ef427ad6085b01e