Set Last-Modified on MarkDown->HTML plugin pages
When a Plugin MarkDown file is served and transparently
converted into an HTML page, the HTTP Last-Modified
field was set on the cached resource but *NOT* on
the returned HTTP response.

This was causing a second HTTP 200 even if the resource
was actually not modified from the first request.

With this fix, the second response to the browser
will be an HTTP 304 (not modified) as it should be.

Change-Id: I9304bde5f9633b2b2f965e9d632303fd022d6651

Former-commit-id: ba05a8a2637d3e646599755227b3effc8b49e10d