Fix possible infinite loop in ScheduleConfig
A missing interval key in the Config will cause an infinite loop when
computing the initialDelay as a while loop will keep adding a negative
value (the MISSING_CONFIG = -1) to the startTime expecting it to move in
the future until it passes the now. However, it will keep moving in the
past.

Don't even try to compute the initialDelay when the interval is missing.

Change-Id: Ia18d8709a47cc53f445a076ed873efe6b797710b

Former-commit-id: 4f259705380a35bfe34f6f4353c72207ca05b556