ChangeInserter: fix comparison of Change.Id to Change.Key
Change key types and method names are confusing, and Java's
Object.equals() was not helping us here. As a result we actually ended
up inserting the same ChangeMessage multiple times, which depending on
the gwtorm backend might or might not have succeeded.

This combined with I5d7fb4cc had the interesting consequence of racing
the row version update with the background mergeability flag update.
If the row version update won, it would have the effect of erasing the
result of the mergeability check, causing spurious "Cannot merge"
warnings.

Change-Id: I8aeb0c87bb2475fa9b49685d27ded48059295547

Former-commit-id: 6761264d0848163ab48695e2edcb661a74137d9f