Require 'Administrate Server' capability to see server summary
The show-caches SSH command now only prints out the summary
information if the caller is an administrator. This is done because
the summary information contains system information which must not be
exposed to non-administrators. This means that the behaviour is
changed and users that only have the 'View Caches' capability are not
anymore able to see the summary information as they did before.

Change-Id: I3b5433b456d3b5f9d5de872a731470d2fe20623f
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 4279d4b9c3ac9289cbf467caca43c6d5dcf6dd80