Improve message when removing a reviewer
When a reviewer is removed from a change also the votes of this
reviewer are deleted. This action is recorded as a change message:

  Removed the following approvals:

  * Code-Review-2 by John Doe <john.doe@example.com>

The term 'approval' is often associated with the highest vote in a
label which is needed to make the change submittable. When the votes
of a reviewer are removed this is mostly about negative veto votes.
This is why using 'approvals' in this context is confusing to users
and the message is now changed to say 'votes' instead.

Change-Id: I388f219d151a431ceeef497dd3d1b220f56e7b00
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: ee06a49d7d0f8f9def8907359d1f0ace41cc669c