Let kill SSH command only kill tasks that are visible to the caller
Before this change having the 'Kill Task' capability was sufficient to
kill any task. Now it only allows to kill tasks that are visible to
the user (e.g. tasks for projects that are visible to the user). To
make all tasks visible to a user the 'View Queue' capability has to be
assigned in addition.

This behaviour is now consistent with the kill task REST API.

Change-Id: If1466b4e9025dc219429f28eb0e204649cef4650
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 8917a04ade2b3bb22b5e4bea08343677e2551720