Add a dropdown to control email notifications on user preference page
Add a dropdown 'Email Notifications' in the user preferences to allow
users to control their email notifications.

On selecting 'Disabled' user will not receive any email notifications
from Gerrit.

Remove the checkbox 'CC Me On Comments I write' and add it to the new
dropdown. Functionality remains the same as the original checkbox.

Options in the dropdown are:

1. Enabled
2. CC On Own Comments
3. Disabled

Feature: Issue 989
Change-Id: If40d0d34f39415d5c64ccaa5803fc6321bddd322

Former-commit-id: 207fa6e5db61f57edd54b4085408efd90d15be22