Fix JS resource loading for JsPlugin type (was returning 404)
Because of the change Ie8d9a9ac1 WrappedRequest
getPathInfo() does not include anymore the plugin
name. This broke the JsPlugin support as any attempt
to serve the actual .js file was not able to locate
the plugin name in the path info.

We need now to get the entire request-uri and assert
the format /plugins/<plugin>/static/<plugins.js>

Change-Id: I5d6d434453824d2539a364a2d6c69b687edbbee0

Former-commit-id: 0a3b513050bae46e0ec2b83d85970c0b50dd752f