Disable Gitweb for in-memory acceptance tests
Acceptance tests run in-memory if the @UseLocalDisk annotation is not
used and therefore, there is no gerrit_site. If Gitweb is installed,
it will be automatically configured and its configuration file will be
written in the gerrit_site/tmp folder. Disable Gitweb for in-memory
acceptance tests otherwise we end up with a UNIT_TEST_GERRIT_SITE/tmp
folder that is not cleaned up after the tests.

Change-Id: Ie3e55313cb8e9e4af53b0fb6f12f6bb6a80f1d7a
Former-commit-id: c99c5d9c598a460abdd38884cbc6f12d32873f43