Update description and error messages in review command
The description and error messages are outdated and are legacy of
the command originally being named "approve".

Change them to be more appropriate to the current usage of the
command, which is to apply reviews, which may or may not include
approvals, to patch sets.

Also rename the `approveOne` method to something more appropriate.

Change-Id: If95a24b8998bbac9a2ea410ecc05974674070e06

Former-commit-id: 77c055e61849355e73d79e44de2ad99efcb37713