Fix showing information title on label cells in change screen
When the "Display in Review Category" option is disabled, the label
cells in the change table are supposed to show a title/tooltip with
the name of the label and reviewer:

  Code-Review by David Pursehouse

However since I8858147c this does not work any more.

Restore this functionality.  Also make it show the tooltip when the
setting is "Show Abbreviated Name".

Change-Id: I117ebb4f09c845a642a068fc8392dd00d74bdb4c

Former-commit-id: 7f3a44114767ce298fee6caba9640e17af9abf82