Modify CommentsInNotesUtil to handle removing a notes object
Previously, we were using the inserter to insert null if the list of
comments to write to the NoteMap was empty. However, in this case, we
should be removing the note that existed before. Therefore, I made an
additional method for the purpose of removing the note if we deleted
all remaining comments from that note, removeNoteFromNoteMap().

Callers should call writeCommentsToNoteMap if they have a non-empty
list of comments that they want to write to a note. If they have an
empty list, they should call removeNoteFromNoteMap which requires a
parameter for the RevId for the note where the comments used to be.

Change-Id: Ia91d8698206ab4814039a47190319838ecb541cc

Former-commit-id: 9733ae83b37bb774f897c1b5ae00da093975aebd