Remove age operator when drilling down from a dashboard to a query
By clicking on the section header in a custom dashboard one can
navigate to the query of that section. In a dashboard space is limited
which is why the section queries are often limited by using the limit
and age operators so that a section cannot grow too big. Sometimes it
makes sense to drill down into a query and look at the complete query
result. This is why any limit operator is automatically removed when
drilling down into a section query. Do the same for the age operator
and remove it automatically when drilling down into a section query.

Change-Id: I209be9516e0462a4d04fbba1b3ec128c0d411e7f
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 780ab208e196a7c554856b9f641fcc685c9c44dc