gerrit_plugin: Allow to co-exist in tree and standalone plugins
Expose variable for plugin API that can be used from plugin build. The
path to plugin API artifacts is different between in tree and standalone
mode. To support the build that works for both mode, the plugin API vars
can be used to reference the plugin API. That way this works for both
modes: in gerrit tree and standalone.

Change-Id: Ia6cf1bad769f339be815b81ab528c8ca774e2f85

Former-commit-id: c563e98d9b83a9595035d715d86d1ee41b153611