Expose the history of all reviewers ever on a change
For the purpose of deleting a draft patch-set and all of its comments,
we need a list of every user who has ever been a reviewer on that
change because those are all the users who could possibly have a draft
comment on this change. With this information, we will be able to scan
the draft refs for all those users to ensure that we delete every
comment for this draft patch-set.

Change-Id: I86a8c5242d62991c3c885911f37438e4f422b0da

Former-commit-id: 3d60276a2b0260638079cef5277075076cff4cfb