Let RefControlTest be responsible for creating own test projects
All test cases which test inheritance of the access rights assume
existence of two projects: "local" and "parent", where "parent"
is the parent of the "local".

The ca3f02638753f2ab7d045800e4b741068c639fab [formerly cc00e0056991355b8780bebd221acf3baf1f55e7] moved the creation of
the "parent" into the Util class and it wasn't intuitive to follow
what exactly the "util.getParentConfig()" meant.

Remove the creation of the "parent" project from the Util class and
create the "parent" in the RefControlTest. Also, make sure that "parent"
is set to be the parent of the "local" in the setUp method.

Change-Id: I991a9bc25ca7609b5d9e15e54167446c7b0aa6cf

Former-commit-id: 64750bd4b198991709eb4604898f0312fcb3126d