Limit max number of changes pushed in a batch.
'receive.maxBatchChanges' is the maximum number of changes that
Gerrit allows to be pushed in a batch for review. When it is
exceeded Gerrit rejects the push with an error message.

This setting can be used to prevent users from uploading large
number of changes by mistake, e.g. when pushing deliveries from
external git repositories.

Default is zero, no limit.

Change-Id: Ib9821190d3c88ac9a2efbb552e21c257165a9443

Former-commit-id: e13b65d8d58364148540bd273c6e983ea2bc8d9d