Update Buck
Buck no longer accepts duplicate deps. One test listed a library
that is always included for tests and can be safely dropped.

Buck does not accept duplicate resources, which happens with
gwt_module() loading in srcs, resources and gwt_xml. Convert
to a set and back to a list to remove duplicates transparently.

Change-Id: I0e450ef376040e8fa315ca17ab62a5cebbaf1d67

Former-commit-id: 3bfca9ed0e0a7b180b7458f43c04020d04c98716