Be more efficient about loading banned commits map
We always have a RevWalk available at call sites, so reuse that.

Don't reload the notes on every call to the validator; there may be
many such calls when used from ReceiveCommits. Since ReceiveCommits
already has a NoteMap available, just plumb that down.

Note that we can't simply remove the NoteMap from ReceiveCommits, as
the absence of any banned commits is one of the things used to tell
whether we can short-circuit all commit validation.

Change-Id: Ib8ba2360c45c106fd1ded54ea392bfcb8d98fff7

Former-commit-id: b5db58b4d93b6eb378ec48d1cec6c93fee2b6d44