Fix JS plugin load when using non-root Gerrit URLs
When Gerrit is not on the root URL path the JS plugins
failed to load because of the exact matching required
on the request URL.

Current check is now more flexible by checking only 
the suffix of the JS plugin path.

NOTE: In theory this may recognise even different URLs
than the intended one (e.g. /gerrit/bla/bla/bla/plugins/...)
but however they would not land on the HttpPluginServlet
anyway. That's why "endsWith" is good enough for this check.

Change-Id: I299b244a9fcdc9b00175e2054a7ba7363620979b
Former-commit-id: 90bef326205ac9753a711a844ca488a2a846b48e