VisibleRefFilterIT: Enable notedb
The set of tests and the expected value is almost exactly identical in
the notedb/non-notedb case except for the presence of meta refs. Hack
this into the assert method so we write the notedb-enabled version of
tests, and meta refs are stripped from the expected sets in this case.
This is fairly hacky, but avoids duplication of the test methods
themselves (or, worse, forgetting to duplicate them.)

Change-Id: I8bb071c920aa2c5be7b4d328036af735a020e1e1

Former-commit-id: 62d0a15324f6627dd685c0bf55537dd3651a73f8