Load SecureStore implementation from site lib dir
This approach shares core ideas of the servlet filter and is alternative
approach to secure store plugin.

To use this extension additional jar file needs to be put in
$site_path/lib directory and gerrit.secureStoreClass property needs to
be added to gerrit.config file. This property must contain fully
qualified class name that implements
com.google.gerrit.server.securestore.SecureStore interface.

During Gerrit start up the class configured with the
gerrit.secureStoreClass will be loaded and the SecureStore interface
will be bound to it.

Change-Id: I90972d28ecab0818d782c3d43acb8c70ba598d12
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>

Former-commit-id: 9927a96b0d60717d0a289c85569ac0d09a8d7ee9