Delete GWT DevMode launcher
There is now only one supported way to debug GWT, so remove the "SDM"
naming, as it's basically an implementation detail. Be consistent
between the launcher and class names,
gerrit_gwt_dbg -> GerritGwtDebugLauncher.

Update the documentation to be consistent as well, and link to the
article about Super Dev Mode.

Change-Id: If541bc48bf828a8af64d631d7acd6f663d5ecea1

Former-commit-id: 9c4a47051f6c1f7e9af6e229cb78cacc870d3584