Allow debug and trace log statements to be printed in log files
System logs file appenders were created with an info threshold which
was preventing debug and trace log statement to be printed in the
files. File appenders should not have a threshold. Instead, the root
logger should have a proper level (i.e. info) so if a specific logger is
using a finer grain level (e.g. debug, trace), the log statements won't
be filtered out by the appender.

Set the root category to info and no longer set the file appender
threshold when creating system logs.

Change-Id: I747bef1f5dcb99784ba91dba6d29cf1fba582b62

Former-commit-id: 0cbe223e404190a1f80e7f2ebbf0b33dad2e94f8