Watchman: Add buck-out dir to ignored directories
As pointed out in [1] and documented in [2] add .watchmanconfig file
to ignore Buck build directory.

This is useful to ignore a directory that contains only build products
and where file change notifications are unwanted because of the sheer
volume of files.  Not to mention that during tests execution temporary
directories are created and wiped out that may confuse watchman.

[1] https://github.com/facebook/buck/issues/183
[2] https://facebook.github.io/watchman/docs/config.html

Change-Id: Ib5a248f9b14006c64fb55d064d8ef5decf6b9061

Former-commit-id: 7cfc17bca7f7be06972dd724088a3be1ba4b2635