ChangeScreen: Highlight selected outdated patch set
Gives a visual hint when the currently viewed patch set is not the
latest one.

An example use-case is when change A depends on some previous patch set
of change B. Clicking change B in the Related Changes tab opens the
dependent patch set, but usually the latest PS is what interests the
user.

The color used for this indication is the same as the marker color for
not current patch sets in Related Changes.

Change-Id: If7d1e673172a44ba1c5ef9aea332d8fc54b1125c

Former-commit-id: ce43e0f7dd885d71153b8a2b574311bf1169d952