Hashtag REST API endpoints
This change introduces three endpoints to ADD/DELETE/GET hashtags
from/to a change.

Whenever the hashtags are successfully updated, the reply will
describe the new complete set of hashtags for that change.

Please note that these endpoints will not be available without
enabling the NoteDB (later on this is expected to be enabled
per default).

Examples:

  PUT /a/changes/1/hashtags HTTP/1.1
  { "hashtags" : "tag1,tag2,tag3" }

  Response:
  HTTP/1.1 200 OK
  [
    "tag1",
    "tag2",
    "tag3"
  ]

  DELETE /a/changes/1/hashtags HTTP/1.1
  { "hashtags" : "tag1" }

  Response:
  HTTP/1.1 200 OK
  [
    "tag2",
    "tag3"
  ]

  GET /a/changes/1/hashtags HTTP/1.1

  Response:
  HTTP/1.1 200 OK
  [
    "tag2",
    "tag3"
  ]

Change-Id: I79dce433e79c6d209268fd434b2311aa551ffc09

Former-commit-id: 34d0db74ef96366117700bb7745e9ff7d76e04e6