Add API to make SecureStore switching possible
In order to be able to implement switching the SecureStore
implementation we need to obtain a list of the currently stored values in the
SecureStore. For this purpose a list() method was added to SecureStore
interface.

Then we can decode them using the old implementation and encrypt with new
one and also replace the secure store jar file in the lib folder and
change the configuration in the gerrit.config in a separate site
program.

Change-Id: Ia270d0ee4a708b5724601902f3382bea3ae45356
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>

Former-commit-id: d984b9c6ce72f3025b5ff5ba6e791b512460de8f