Avoid CodeReviewCommit.error() whenever possible
The instances returned by this static factory method are severely
limited, in that they are not tied to actual commit objects. In most
cases, we can tie them to commit objects, so do that whenever
possible. Limit the statuses that we can pass to this method by
making it private, and creating separate factories only for those
statuses where there is really no commit.

Change-Id: I0cb392f4f7f5134d4a7487a2a7bb7ac9d770f334

Former-commit-id: 002b9ee7b3dd78489edcfed8e0769ecb4a41aa9f