Add diff pref whether the diff table header should be auto hidden
Add a new diff preference that allows the user to decide whether the
diff table header with the patch set selection should be automatically
hidden when scrolling down more than half of a page. At the moment
this behaviour cannot be disabled, but some users find it annoying
that the header disappears and that they need to scroll up to be able
to change the patch set selection.

By default the diff table header is automatically hidden as this is
the current behaviour.

Change-Id: I1d65c5058222c924ccaf3276f7d2d83fac7f4d32
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 1b6c6d843c36b899eee883dda2bfa7714998a7fc