Improve display of long user names for collapsed comments in history
The width of the column in which user names for collapsed comments are
shown in the history on the change screen is fixed so that long user
names get truncated, e.g. 'AutomaticGerritVoterLinux' is shown as
'AutomaticGerritVo...'. This is bad if there are several users with
long user names that have the same prefix, e.g.
'AutomaticGerritVoterWindows' would be also shown as
'AutomaticGerritVo...' and a user must expand the comment to see the
full user name.

Improve this by inserting the ellipsis in the middle of the user name
so that the start and end of the user name are always visible, e.g.
"AutomaticG...VoterLinux" and "AutomaticG...terWindows".

Also slightly increase the width of the column that is used for user
names of collapsed comments.

Bug: issue 2837
Change-Id: I995a848381a31b44f60e018fb1173351766df587
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 8dc6bd3179083f79fc1c7eb71017d03b2bdbd088