Enable inline edit in side by side view
Switch CodeMirror to edit mode in "B" side to enable inline edit.

In edit mode, CodeMirror is switched to "default" mode because "vim"
mode was sometimes not responsive. Also commenting is disabled in this
mode.

During testing I've discovered a small bug in CodeMirror. A JS
exception is thrown after entering a text and going to the new line.
Everything looks and works fine afterwards, but an error stays in the
JS console. Therefore it was suppressed in JSNI code.

Edit icon is taken from the Tango project:
  http://tango.freedesktop.org/Tango_Icon_Library
Specifically, 16x16/devices/media-floppy.png from
  http://tango.freedesktop.org/releases/tango-icon-theme-0.8.90.tar.gz

Change-Id: Ie9b2c5f2b362da5ffbead57a241756a45e9eb60c
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>

Former-commit-id: 00356c2c78957ff38e48dcb9f76adffe038b0ded