Hashtags: Refactoring of the REST API
Sending data with a DELETE is not a common standard. Replace the PUT and
DELETE endpoints with a POST method which accepts both adds and removals.

Instead of sending hashtags as a comma-separated list, send as a set and
raise HTTP 400 Bad Request if any hashtag contains a comma.

Return results as an ordered list. Having the results in a deterministic
order makes it easier to test.

Only perform notes db update and indexing if any tags were actually added
or removed.

Examples:

  GET /a/changes/1/hashtags HTTP/1.1

  Response:
  HTTP/1.1 200 OK
  [
    "tag1",
    "tag2"
  ]

  POST /a/changes/1/hashtags HTTP/1.1
  { "add" : "tag3", "remove" : "tag2" }

  Response:
  HTTP/1.1 200 OK
  [
    "tag1",
    "tag3"
  ]

Change-Id: Idde97697ecc6bc1a51404bef67baf645d8909555

Former-commit-id: 6b15344e38f9721884149e11865166b9049d908c