Fix SecureStore binding for plugins
Some plugins like uploadvalidator, will now not load because of missing
binding for SecureStore in CopyConfigModule.

To be able to bind SecureStoreProvider there we need to move it from pgm
module to server. Then because it is using IoUtils and
SiteLibraryLoaderUtil to common module and adjust BUCK configuration to
make it compile Gerrit.

Change-Id: I20d7f56750e5290872b984881cd4dbd1add5988c
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>
Former-commit-id: 7f4662fd4a3e02fc0c1a50181166520c985bfa1c