Fix TopMenu extensions for the Project Menu items
The behaviour of TopMenu extensions to the TopMenu projects menubar
was broken after the extension of “Project-aware” menu items I516f129dc
Even project agnostic items were incorrectly processed as
project-specific causing them to be hidden when there was no active
project selected.

The issue has been discussed on the mailing list (see [1]) and 
acknowledged to be a regression to be fixed. Issue can be easily 
reproduced looking at the cookbook-plugin: “Browse Repositories” 
TopMenu entry is hidden whilst it should be normally shown.
[1] https://groups.google.com/forum/#!topic/repo-discuss/nFBQFc0MS6E

Change-Id: I8adb59ddda3cae6fc5819c525580096002377c5b
Former-commit-id: b9c8a1e40c71386b6811707c7cdf3610d13ab03c