Clarify the change REST API documentation.
The review REST API doesn't seem to work as documented.  The
documentation says that the GET review command should return
a ChangeInfo object which contains a RevisionInfo object
and the RevisionInfo object should contain weblinks, files,
actions, etc. This is a true statement although it isn't
complete.  The missing info is that the user must add special
fields to the request to get the additional RevisionInfo.
Hopefully this change will clarify the usage of this REST command.

Bug: issue 2913
Change-Id: I79f67c7aa986cc9a2fc16da9f8d81d6ec437b14a

Former-commit-id: b3139b753b4afed150552167e5de0edd5a48ce5e