Remove mergeable bit from Change
Upgrade the schema to remove the field, but do not actually migrate
data. This will be backfilled done during the next site reindex.

The reason for this suboptimal upgrade process is that it would be
nontrivial to copy the data in a schema upgrade, since the key for the
cache requires running the entire submit rule evaluator to get the
submit type. The dbInjector used to create the chain of SchemaVersions
has far too few bindings for this. Getting these at schema upgrade
time would require a completely different approach to injecting schema
upgrade steps.

Change-Id: I6fd6de3d06ed92ee4d2dbea43f3e9974e9491542

Former-commit-id: 1c8767ed1407a910a1b48ac7b46ed617da61d06a