Cleanup test of ScheduleConfig
Asserting not equals to missing doesn't sufficiently test the
value was parsed properly for a non-standard key. Instead check
for real values.

Replace the hacky MessagseFormat string parsing of Config with
API setters. The Config class is already well tested in JGit.
There is no need to obfuscate this code to "verify" again the
Config object can read from a string.

Change-Id: I60cbf6dae9621fe7a7d7a10c5db87a39d3b57aac

Former-commit-id: a19dbb5f9aded2a502c9e2115977aca7f80d2e62